project4_RAG/requirements.txt
streamlit>=1.25.0
openai>=1.0.0
pinecone>=3.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
tqdm>=4.0.0 